% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccoev.R
\name{cc}
\alias{cc}
\title{Cooperative coevolution.}
\usage{
cc(population, fun, group = NULL)
}
\arguments{
\item{population}{Initial population}

\item{fun}{The objective function to be solved}

\item{group}{Vector of list. Each list contains a group of non-separable variables.}
}
\description{
Main function for the cooperative coevolution procedure.
}
