% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_2.R
\name{cc_2}
\alias{cc_2}
\title{Cooperative coevolution.}
\usage{
cc_2(contextVector = NULL, nVar, fun, budget = 1e+06, group = NULL,
  grouping_control = list(), nCycle = 9, lbound = rep(-Inf, nVar),
  ubound = rep(Inf, nVar), evalInterval = 1e+05, ...)
}
\arguments{
\item{fun}{The objective function object to be solved}

\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{differential_grouping}.}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}

\item{...}{Further arguments passed to \code{fun}.}

\item{population}{Initial population}

\item{phaseSolver}{Optimizer/solver for phase-optimization step.}
}
\description{
Main function for the cooperative coevolution procedure.
}
\examples{
population <- (randtoolbox::sobol(5,100))
optimum <- rep(0,100)
func <- f1cec
ctrl <- list(lbound=rep(-100,100),ubound=rep(100,100),delta=rep(100,100))
group <- differential_grouping(ncol(population),func,ctrl,o=optimum)
cc_2(nVar = 100,fun=func,phaseSolver = cmaes,group=group,nCycle = 2,o=optimum,lbound=rep(-100,100),ubound=rep(100,100))
}
