% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC_CMA.R
\name{CC_CMA}
\alias{CC_CMA}
\title{Cooperative coevolution.}
\usage{
CC_CMA(contextVector = NULL, nVar, fun, group = NULL, budget = 1e+06,
  lbound = rep(-Inf, nVar), ubound = rep(Inf, nVar), nLevel = 4,
  evalInterval = 1e+05, SA_method = "morris", ...)
}
\arguments{
\item{fun}{The objective function object to be solved}

\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{differential_grouping}.}

\item{...}{Further arguments passed to \code{fun}.}

\item{population}{Initial population}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}
}
\description{
Main function for the cooperative coevolution MOFBVE with multilevel framework
}
\examples{
optimum <- rep(13,1000)
func <- f1cec
ctrl <- list(lbound=rep(-100,1000),ubound=rep(100,1000),delta=rep(20,1000))
SACC(nVar = 1000,fun=func,lbound=rep(-100,1000),ubound=rep(100,1000),o=optimum)
}
