% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccoev_ex.R
\name{cc_ex}
\alias{cc_ex}
\title{Cooperative coevolution.}
\usage{
cc_ex(contextVector = NULL, nVar, fun, phaseSolver = cmaes,
  budget = 1e+06, lbound = rep(-Inf, length(population)),
  ubound = rep(Inf, length(population)), nLowerLevel = round(0.1 *
  nVar), ...)
}
\arguments{
\item{fun}{The objective function object to be solved}

\item{phaseSolver}{Optimizer/solver for phase-optimization step.}

\item{...}{Further arguments passed to \code{fun}.}

\item{population}{Initial population}

\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{differential_grouping}.}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}
}
\description{
Main function for the cooperative coevolution procedure.
}
\examples{
optimum <- rep(13,1000)
func <- f1cec
ctrl <- list(lbound=rep(-100,1000),ubound=rep(100,1000),delta=rep(20,1000))
cc_ex(nVar = 1000,fun=func,phaseSolver = cmaes,lbound=rep(-100,1000),ubound=rep(100,1000),o=optimum)
}
