% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ccoevolution-package}
\alias{ccoevolution-package}
\alias{ccoevolution}
\title{Cooperative Coevolution for LSGO}
\description{
This package is an implementation of cooperative coevolution for large scale global optimization.
To use this package, several Python modules must be installed beforehand, namely: numpy and PyGMO.
}
\details{
Note: This package uses column-major ordering, i.e. an individual should be contained in a single column, each row represents different variable.

\tabular{ll}{
Package: \tab ccoevolution\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2019-07-16\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\section{Acknowledgments}{

This work is funded by the European Commission's H2020 programme through
the UTOPIAE  Marie  Curie  Innovative  Training Network, H2020-MSCA-ITN-2016,
under Grant Agreement No. 722734.
}

\section{Maintainer}{

Dani Irawan \email{irawan_dani@yahoo.com}
}

\seealso{
Main interface function is \code{\link{ccoev}}.
}
\author{
Dani Irawan \email{irawan_dani@yahoo.com}
}
\keyword{package}
