% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_es_with_IPOP.R
\name{cma_es}
\alias{cma_es}
\title{CMA-ES}
\source{
The code is a modification from cmaes::cma_es()
}
\usage{
cma_es(
  par,
  fn,
  ...,
  lower,
  upper,
  control = list(),
  inputScaleFactor = rep(1, length(par)),
  inputScaleShift = rep(0, length(par)),
  repair = F
)
}
\arguments{
\item{par}{Initial mean value for search. Do not scale.}

\item{inputScaleFactor}{Multiplier for scaling.}

\item{inputScaleShift}{The shift for scaling.}
}
\description{
Global optimization procedure using a covariance matrix adapting
evolutionary strategy. A scaling factor for inputs can be used.
The input will be processed as: \code{(par-inputScaleShift)*inputScaleFactor}.
}
\seealso{
\code{\link[cmaes]{cma_es}}
}
