% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_es_bilevel.R
\name{cma_es_bilevel}
\alias{cma_es_bilevel}
\title{Global optimization procedure using a covariance matrix adapting
evolutionary strategy for separable functions.}
\source{
The code is a minor modification from cmaes::cma_es(), adding cutoff on the box constraint. (line 128)
}
\usage{
cma_es_bilevel(par, fn, ..., lower, upper, control = list())
}
\description{
Global optimization procedure using a covariance matrix adapting
evolutionary strategy for separable functions.
}
\seealso{
\code{\link[cmaes]{cma_es}}
}
