% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{cmaes}
\alias{cmaes}
\title{Phase optimization using (mu+mu)-CMA-ES}
\usage{
cmaes(contextVector, func, groupMember, ..., ubound = rep(1,
  length(groupMember)), lbound = rep(0, length(groupMember)),
  popSize = length(groupMember), maxit = 10000)
}
\arguments{
\item{contextVector}{The current best solution}

\item{groupMember}{Vector containing the current group member, i.e., which variables among the \code{contextVector} to be optimized in the current phase.}

\item{...}{Further arguments passed to \code{fun}.}

\item{ubound}{upper bound for the group}

\item{lbound}{lower bound for the group}

\item{popSize}{population size.}

\item{fun}{function to be optimized}
}
\value{
Returns a list containing x (decision variable) and y (objective value)
}
\description{
Wrapper function, the optimization itself is using the \code{cmaes::cma_es()} function.
}
