% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{cmaes_1}
\alias{cmaes_1}
\title{Phase optimization using (1+1)-CMA-ES}
\usage{
cmaes_1(contextVector, func, groupMember, ..., ubound = 1, lbound = 0,
  popSize = 1)
}
\arguments{
\item{contextVector}{The current best solution}

\item{groupMember}{Vector containing the current group member, i.e., which variables among the \code{contextVector} to be optimized in the current phase.}

\item{...}{Further arguments passed to \code{fun}.}

\item{ubound}{upper bound for the group}

\item{lbound}{lower bound for the group}

\item{popSize}{population size. Ignored in cmaes_1}

\item{fun}{function to be optimized}
}
\value{
Returns a list containing x (decision variable) and y (objective value)
}
\description{
Wrapper function, the optimization itself is using the \code{cmaes::cma_es()} function.
Use this function as a template to create a \code{phaseSolver} with single individual.
}
