% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_grouping.R
\name{DG2}
\alias{DG2}
\title{Differential Grouping version 2}
\usage{
DG2(nVar, fun, control = list(), ...)
}
\arguments{
\item{nVar}{Number of variables to be grouped}

\item{fun}{The objective function to be solved}

\item{control}{List of control parameter for DG
\code{lbound} A vector of lower bound values for each variable
\code{ubound} A vector of upper bound values for each variable
\code{delta} The shift from \code{lbound} to compute the gradient}

\item{...}{Further arguments to be passed to fun}
}
\value{
A list
        \code{group} Lists of the groups found by DG2
        \code{separable} Vector of the separable variables found by DG2
        \code{DSM} The full design structure matrix indicating the interaction graph.
        \code{nEval} number of evaluation used in grouping
}
\description{
Create groups of variable using Differential Grouping. The code uses the supremum tolerance.
Infimum tolerance is calculated, but given zero weight.
}
\examples{
control <- list(lbound=rep(-100,100),ubound=rep(100,100),delta=0.1)
groups <- DG2(nVar=100,f1cec,control,o=rep(13,100))
}
\references{
<doi:10.1109/TEVC.2017.2694221>
}
