% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_grouping.R
\name{differential_grouping}
\alias{differential_grouping}
\title{Differential Grouping}
\usage{
differential_grouping(nVar, fun, lbound = rep(0, nVar), ubound = rep(1,
  nVar), delta = NULL, tolerance = 0.001, ...)
}
\arguments{
\item{nVar}{Number of variables to be grouped}

\item{fun}{The objective function to be solved}

\item{lbound}{A vector of lower bound values for each variable}

\item{ubound}{A vector of upper bound values for each variable}

\item{delta}{The shift from \code{lbound} to compute the gradient}

\item{tolerance}{Tolerance \epsilon indicates the threshold on the objective function to determine whether two variables are separable or not. Smaller tolerance will make it easier for variables to be grouped together.}

\item{...}{Further arguments to be passed to fun}
}
\description{
Create groups of variable using differential grouping
}
\references{
Omidvar, et al.
}
