% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgx.R
\name{differential_grouping_ex}
\alias{differential_grouping_ex}
\title{Differential Grouping}
\usage{
differential_grouping_ex(nVar, fun, control = list(), ...)
}
\arguments{
\item{nVar}{Number of variables to be grouped}

\item{fun}{The objective function to be solved}

\item{control}{List of control parameter for DG
\code{lbound} A vector of lower bound values for each variable
\code{ubound} A vector of upper bound values for each variable
\code{delta} The shift from \code{lbound} to compute the gradient
\code{tolerance} Tolerance indicates the threshold on the objective function to determine whether two variables are separable or not. Smaller tolerance will make it easier for variables to be grouped together.}

\item{...}{Further arguments to be passed to fun}
}
\description{
Create groups of variable using differential grouping
}
\examples{
control <- list(lbound=-100,ubound=100,delta=0.1)
groups <- differential_grouping_ex(100,f1cec,control)
}
\references{
Omidvar, et al.
}
