% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f10cec_s}
\alias{f10cec_s}
\title{Rotated Ackley function with 1000D variables. no separable variable}
\usage{
f10cec_s(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f10 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f10cec_s(x,o)
}
