% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f11cec_s}
\alias{f11cec_s}
\title{Rotated Schwefel function with 1000D variables. no separable variable}
\usage{
f11cec_s(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f11 function.
}
\examples{
x <- runif(100)
o <- double(100)
f11cec_s(x,o)
}
