% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f12cec_s}
\alias{f12cec_s}
\title{Shifted Rosenbrock function with 1000D variables. no separable variable}
\usage{
f12cec_s(x, o = NULL)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f12 function.
}
\examples{
x <- runif(100)
o <- double(100)
f12cec_s(x,o)
}
