% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f13cec_s}
\alias{f13cec_s}
\title{Schwefel function with Conforming Overlapping Subcomponents}
\usage{
f13cec_s(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f13 function.
}
\examples{
x <- runif(94)
o <- double(94)
f13cec_s(x,o)
}
