% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec_vectorized.R
\name{f15cec}
\alias{f15cec}
\title{shifted Schwefel function}
\usage{
f15cec(x, o = NULL, seed = 1000)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f15 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f15cec(x,o)
}
