% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f15cec_s}
\alias{f15cec_s}
\title{shifted Schwefel function}
\usage{
f15cec_s(x, o = NULL)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f15 function.
}
\examples{
x <- runif(100)
o <- double(100)
f15cec_s(x,o)
}
