% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec.R
\name{f1cec}
\alias{f1cec}
\title{Shifted elliptic function.}
\usage{
f1cec(x, o = NULL)
}
\arguments{
\item{x}{design variable. One row per point.}

\item{o}{the optimum}
}
\description{
CEC 2013 f1 function. Can accomodate different variable dimensionality m>0. The CEC2013 test problem uses m=1000.
}
\examples{
x <- runif(1000)
o <- double(1000)
f1cec(x,o)
}
