% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec.R
\name{f2cec}
\alias{f2cec}
\title{Shifted Rastrigin's function with m>0 Dimensions variables. The CEC2013 test problem uses m=1000}
\usage{
f2cec(x, o = NULL)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}
}
\description{
CEC 2013 f2 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f2cec(x,o)
}
