% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec_vectorized.R
\name{f3cec}
\alias{f3cec}
\title{Shifted Ackley's function with m>0 variables. The CEC2013 test problem uses m=1000.}
\usage{
f3cec(x, o = NULL)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}
}
\description{
CEC 2010 f3 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f3cec(x,o)
}
