% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec.R
\name{f4cec}
\alias{f4cec}
\title{Rotated elliptic function with 1000D variables.}
\usage{
f4cec(x, o, seed = 2000)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f4 function. The permutation can be modified by setting the seed. The problem is fixed to have 1000D.
}
\examples{
x <- runif(1000)
o <- double(1000)
f4cec(x,o)
}
