% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f4cec_s}
\alias{f4cec_s}
\title{Rotated elliptic function with 1000D variables.}
\usage{
f4cec_s(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f4 function. The permutation can be modified by setting the seed. The problem is fixed to have 1000D.
}
\examples{
x <- runif(100)
o <- double(100)
f4cec_s(x,o)
}
