% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_scaled_down.R
\name{f5cec_s}
\alias{f5cec_s}
\title{Rotated rastrigin function with 1000D variables.}
\usage{
f5cec_s(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f5 function.
}
\examples{
x <- runif(100)
o <- double(100)
f5cec_s(x,o)
}
