% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec.R
\name{f6cec}
\alias{f6cec}
\title{Rotated Ackley function with 1000D variables.}
\usage{
f6cec(x, o, seed = 1000)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f6 function.
}
\examples{
x <- runif(100)
o <- double(100)
f6cec(x,o)
}
