% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec.R
\name{f7cec}
\alias{f7cec}
\title{Rotated Schwefel function with 1000D variables.}
\usage{
f7cec(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f7 function.
}
\examples{
x <- runif(100)
o <- double(100)
f7cec(x,o)
}
