% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec_vectorized.R
\name{f9cec}
\alias{f9cec}
\title{Rotated rastrigin function with 1000D variables. no separable variable}
\usage{
f9cec(x, o = NULL, seed = 1000)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f9 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f9cec(x,o)
}
