% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_level_large_scale.R
\name{ML_LSGO}
\alias{ML_LSGO}
\title{Cooperative coevolution.}
\usage{
ML_LSGO(
  nVar,
  fun,
  ...,
  group = NULL,
  variable_importance = NULL,
  budget = 3e+06,
  lbound = NULL,
  ubound = NULL,
  nLevel = 4,
  eval_interval = 1e+05,
  SA_method = c("morris_mu", "morris_k", "rf", "sobol"),
  maxSeqIter = 0,
  fname_prefix = "datalogf1_",
  infill = c("best", "ei"),
  mode = "optim"
)
}
\arguments{
\item{fun}{The objective function object to be solved}

\item{...}{Further arguments passed to \code{fun}.}

\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{differential_grouping}.}

\item{population}{Initial population}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}
}
\description{
Main function for the cooperative coevolution with multilevel framework.
}
\examples{
optimum <- rep(13,1000)
func <- f1cec
ctrl <- list(lbound=rep(-100,1000),ubound=rep(100,1000),delta=rep(20,1000))
ML_LSGO(nVar = 1000,fun=func,lbound=rep(-100,1000),ubound=rep(100,1000),o=optimum)

}
