% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mofbve_grouped.R
\name{mofbve_grouped}
\alias{mofbve_grouped}
\title{Cooperative coevolution.}
\usage{
mofbve_grouped(contextVector = NULL, nVar, group = NULL, fun,
  phaseSolver = cmaes, budget = 1e+06, lbound = rep(-Inf, nVar),
  ubound = rep(Inf, nVar), nLevel = 4, ...)
}
\arguments{
\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{DG2}.}

\item{fun}{The objective function object to be solved}

\item{phaseSolver}{Optimizer/solver for phase-optimization step.}

\item{...}{Further arguments passed to \code{fun}.}

\item{population}{Initial population}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}
}
\description{
Main function for the cooperative coevolution MOFBVE with multilevel framework
}
\examples{
optimum <- rep(13,1000)
func <- f1cec
ctrl <- list(lbound=rep(-100,1000),ubound=rep(100,1000),delta=rep(20,1000))
mofbve_grouped(nVar = 1000,fun=func,phaseSolver = cmaes,lbound=rep(-100,1000),ubound=rep(100,1000),o=optimum)
}
