% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_CMA_optim.R
\name{optim_sepCMA}
\alias{optim_sepCMA}
\title{Cooperative coevolution.}
\usage{
optim_sepCMA(
  contextVector = NULL,
  nVar,
  fun,
  ...,
  group = NULL,
  budget = 3e+06,
  lbound = rep(-Inf, nVar),
  ubound = rep(Inf, nVar),
  evalInterval = 1e+05
)
}
\arguments{
\item{fun}{The objective function object to be solved}

\item{...}{Further arguments passed to \code{fun}.}

\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{differential_grouping}.}

\item{population}{Initial population}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}
}
\description{
Main function for the cooperative coevolution MOFBVE with multilevel framework
}
\examples{
optimum <- rep(13,1000)
func <- f1cec
ctrl <- list(lbound=rep(-100,1000),ubound=rep(100,1000),delta=rep(20,1000))
optim_sepCMA(nVar = 1000,fun=func,lbound=rep(-100,1000),ubound=rep(100,1000),o=optimum)
}
