% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{optimDE}
\alias{optimDE}
\title{Phase optimization using Differential Evolution}
\usage{
optimDE(contextVector, fun, groupMember, ..., ubound = 1, lbound = 0,
  popSize = length(groupMember))
}
\arguments{
\item{contextVector}{The current best solution}

\item{fun}{function to be optimized}

\item{groupMember}{Vector containing the current group member, i.e., which variables among the \code{contextVector} to be optimized in the current phase.}

\item{...}{Further arguments passed to \code{fun}.}

\item{ubound}{upper bound for the group}

\item{lbound}{lower bound for the group}

\item{popSize}{population size.}
}
\value{
Returns a list containing x (decision variable) and y (objective value)
}
\description{
Wrapper function, the optimization itself is using the \code{DEoptim::DEoptim()} function.
Use this function as a template to create a population-based \code{phaseSolver}.
}
