% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SACC_DE.R
\name{SACC_DE}
\alias{SACC_DE}
\title{Cooperative coevolution.}
\usage{
SACC_DE(contextVector = NULL, nVar, fun, ..., group = NULL,
  budget = 3e+06, lbound = rep(-Inf, nVar), ubound = rep(Inf, nVar),
  nLevel = 4, evalInterval = 1e+05, SA_method = c("morris_mu",
  "morris_k", "rf", "sobol"))
}
\arguments{
\item{fun}{The objective function object to be solved}

\item{...}{Further arguments passed to \code{fun}.}

\item{group}{Vector of list. Each list contains a group of non-separable variables. Available method: \code{differential_grouping}.}

\item{population}{Initial population}

\item{grouping_control}{Grouping parameters. The members depends on which grouping being used.}

\item{nCycle}{Number of cycles to be run.}
}
\description{
Main function for the cooperative coevolution MOFBVE with multilevel framework
}
\examples{
optimum <- rep(13,1000)
func <- f1cec
ctrl <- list(lbound=rep(-100,1000),ubound=rep(100,1000),delta=rep(20,1000))
SACC_DE(nVar = 1000,fun=func,lbound=rep(-100,1000),ubound=rep(100,1000),o=optimum)
}
