% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_grouping.R
\name{XDG}
\alias{XDG}
\title{Extended Differential Grouping}
\usage{
XDG(nVar, fun, control = list(), ...)
}
\arguments{
\item{nVar}{Number of variables to be grouped}

\item{fun}{The objective function to be solved}

\item{control}{List of control parameter for DG
\code{lbound} A vector of lower bound values for each variable
\code{ubound} A vector of upper bound values for each variable
\code{delta} The shift from \code{lbound} to compute the gradient}

\item{...}{Further arguments to be passed to fun}
}
\description{
Create groups of variable using extended differential grouping (XDG). Tolerance is set following DG2 scheme.
}
\examples{
control <- list(lbound=rep(-100,100),ubound=rep(100,100),delta=0.1)
groups <- XDG(nVar=100,f1cec,control,o=rep(13,100))
}
\references{
<doi:10.1145/2739480.2754666>
}
