% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma_es_with_IPOP.R
\name{cma_es}
\alias{cma_es}
\title{Covariance matrix adapting evolutionary strategy for separable problem}
\source{
The code is a minor modification from cmaes::cma_es(), adding cutoff on the box constraint. (line 128)
}
\usage{
cma_es(
  par,
  fn,
  ...,
  lower,
  upper,
  control = list(),
  logFeasible = F,
  limit_sigma = F
)
}
\description{
Global optimization procedure using a covariance matrix adapting
evolutionary strategy for separable functions.
}
\seealso{
\code{\link[cmaes]{cma_es}}
}
