% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec_vectorized.R
\name{f12cec}
\alias{f12cec}
\title{Shifted Rosenbrock function with 1000D variables. no separable variable}
\usage{
f12cec(x, o = NULL, seed = 1000)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f12 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f12cec(x,o)
}
