% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec_vectorized.R
\name{f8cec}
\alias{f8cec}
\title{Groups of nonseparable Rotated elliptic function with 1000D variables.}
\usage{
f8cec(x, o = NULL, seed = 1000)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f8 function.
}
\examples{
x <- runif(1000)
o <- double(1000)
f8cec(x,o)
}
