% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problem_cec.R
\name{f9cec}
\alias{f9cec}
\title{Rotated rastrigin function with 1000D variables. no separable variable}
\usage{
f9cec(x, o = NULL, rotation_matrix, permutation, weight)
}
\arguments{
\item{x}{design variable}

\item{o}{the optimum}

\item{seed}{Seed for random permutation and rotation matrix}
}
\description{
CEC 2013 f9 function.
}
\examples{
x <- runif(100)
o <- double(100)
f9cec(x,o)
}
