% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_cmaes.R
\name{sep_cma_es}
\alias{sep_cma_es}
\title{Covariance matrix adapting evolutionary strategy for separable problem}
\source{
The code is a minor modification from cmaes::cma_es(), specialized to handle separable function
}
\usage{
sep_cma_es(par, fn, ..., lower, upper, control = list())
}
\description{
Global optimization procedure using a covariance matrix adapting
evolutionary strategy for separable functions.
}
\references{
<doi:10.1007/978-3-540-87700-4_30>
}
\seealso{
\code{\link[cmaes]{cma_es}}
}
